requirements.txt
aiohttp==3.5.4
amqp==2.5.1
appdirs==1.4.3
argh==0.26.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.8.0
billiard==3.6.1.0
bs4==0.0.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
chart-studio==1.0.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
curio==0.9
cycler==0.10.0
decorator==4.4.0
dnspython==1.16.0
dramatiq==1.7.0
EasyProcess==0.2.7
eventlet==0.25.1
fake-useragent==0.1.11
gevent==1.4.0
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
kiwisolver==1.1.0
kombu==4.6.4
lxml==4.4.1
matplotlib==3.1.1
monotonic==1.5
more-itertools==7.2.0
multidict==4.5.2
numpy==1.17.2
pandas==0.25.1
parse==1.12.1
parsel==1.5.2
parso==0.5.1
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
pika==1.1.0
plotly==4.1.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
pyee==6.0.0
Pygments==2.4.2
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyppeteer==0.0.25
pyquery==1.4.0
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.2
PyVirtualDisplay==0.2.4
PyYAML==5.1.2
queuelib==1.5.0
redis==3.3.8
requests==2.22.0
requests-html==0.10.0
retrying==1.3.3
Scrapy==1.7.3
scrapy-selenium==0.0.7
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.3
tqdm==4.35.0
traitlets==4.3.2
Twisted==19.7.0
urllib3==1.25.3
vine==1.3.0
w3lib==1.21.0
watchdog==0.8.3
watchdog-gevent==0.1.0
wcwidth==0.1.7
websockets==8.0.2
yarl==1.3.0
zipp==0.6.0
zope.interface==4.6.0
